(CSV-164) CSVParser now keeps track of all empty header indexes if empty headers are allowed.
This is a partial fix for the issue in (CSV-164) as it only helps in the case of empty headers. As this is the only case where duplicate header elements are currently supported at all, it seemed a pretty important one to fix, though.

The header map still tracks the last empty header only for backward compatibility.

There are currently no specific unit tests checking for this behaviour.